Sync `maxFrameRate` with RTCRtpEncodingParameters interface specification https://bugs.webkit.org/show_bug.cgi?id=307249 rdar://169885744
Reviewed by Philippe Normand.

This patch aligns WebKit with Gecko / Firefox, Blink / Chromium and
Web Specification [1]:

As per specification [1], the `maxFrameRate` should be `double` and
not `unsigned long`, this patch fixes it.

[1] https://w3c.github.io/webrtc-pc/#dom-rtcrtpencodingparameters

* Source/WebCore/Modules/mediastream/RTCRtpEncodingParameters.h:
* Source/WebCore/Modules/mediastream/RTCRtpEncodingParameters.idl:
* Source/WebCore/Modules/mediastream/gstreamer/GStreamerWebRTCUtils.cpp:
(WebCore::fromRTCEncodingParameters): Change to G_TYPE_DOUBLE
* Source/WebCore/platform/mediastream/gstreamer/GStreamerVideoRTPPacketizer.cpp:
(WebCore::GStreamerVideoRTPPacketizer::configure const): Change to `double` and removed unneeded static_cast
* LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpParameters-maxFramerate-expected.txt: Progressions

Canonical link: https://commits.webkit.org/307033@main

+++ Source/WebCore/Modules/mediastream/RTCRtpEncodingParameters.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018 Apple Inc. All rights reserved.
+ * Copyright (C) 2018-2026 Apple Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -37,7 +37,7 @@ struct RTCRtpEncodingParameters : RTCRtpCodingParameters {
 
     bool active { true };
     std::optional<unsigned long> maxBitrate;
-    std::optional<unsigned long> maxFramerate;
+    std::optional<double> maxFramerate;
     std::optional<double> scaleResolutionDownBy;
 
     RTCPriorityType priority { RTCPriorityType::Low };
